Some additions regarding floating IPs
First of all, a timeout of 2s that is required and hopefully
sufficient for a new VM to receive a fixed IP. Without a fixed IP, the
allocation of a floating IP will fail.

Secondly, we need to deallocate the floating IP from an instance
before terminating it. Otherwise the floating IP will remain allocated
to the project without any means to release it via the OCCI interface.